pyproject.toml
[project]
name = "adk-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "litellm>=1.80.11",
]
